updated the project list
